campaign-tests-main/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
dnspython==2.3.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
pymongo==4.3.3
PySocks==1.7.1
pytest==7.3.1
pytest-order==1.1.0
python-dotenv==1.0.0
requests==2.30.0
selenium==4.11.2
seletools==1.4.0
self==2020.12.3
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
webdriver-manager==3.8.6
wsproto==1.2.0
